pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "acr-pivacy-attacks"
version = "0.0.1"
requires-python = "~= 3.8.0"

dependencies = [
	"joblib==0.16.0",
	"matplotlib==3.3.2",
	"Keras==2.4.3",
	"neptune_client==0.4.125",
	"neptune_contrib==0.24.8",
	"pandas==1.1.1",
	"tensorflow-cpu==2.4.0",
	"Keras_Preprocessing==1.1.2",
	"numpy==1.19.4",
	"scikit_learn==0.23.2",
	"imbalanced-learn==0.7.0",
	"tikzplotlib==0.9.6",
]

[project.optional-dependencies]
dev = [
	"ipython",
	"python-language-server[all]",
	"pyls-isort",
	"click",
	"Sphinx",
	"coverage",
	"flake8",
	"python-dotenv>=0.5.1",
]
